Refactor container tests
